* Move extractDuration method to util class for better re-usability. * Minor Lint warning fixes
